# backend/openapi/openapi.yaml
...
openapi: 3.1.0
components:
  schemas:
    AnalyzeRequest:
      type: object
      properties:
        gatewayStrategy:
          type: string
        products:
          type: array
          items:
            type: string
        targetClusterId:
          type: string
...
    ApplyRequest:
      type: object
      properties:
        excludedIndexes:
          type: array
          items:
            type: integer
            format: int32
        yamlOverrides:
          type: object
          additionalProperties:
            type: string
    ApplyResult:
      type: object
      properties:
        planId:
          type: string
        applied:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        results:
          type: array
          items:
            $ref: "#/components/schemas/ResourceResult"
        targetClusterId:
          type: string
...
    BulkRevertRequest:
      type: object
      properties:
        planIds:
          type: array
          items:
            type: string
        deleteGateway:
          type: boolean
    BulkRevertResult:
      type: object
      properties:
        totalPlans:
          type: integer
          format: int32
        totalReverted:
          type: integer
          format: int32
        totalFailed:
          type: integer
          format: int32
        planResults:
          type: array
          items:
            $ref: "#/components/schemas/ApplyResult"
...
    ConfirmRegistrationRequest:
      type: object
      properties:
        componentYaml:
          type: string
...
    DriftEntry:
      description: Drift status for a generated resource on the target cluster
      type: object
      properties:
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
          description: "in-sync, missing, or error"
        message:
          type: string
    GeneratedResource:
      type: object
      properties:
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        yaml:
          type: string
...
    ImportExportResponse:
      type: object
      properties:
        importMode:
          type: string
        productCount:
          type: integer
          format: int32
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductSummary"
        manifest:
          $ref: "#/components/schemas/ManifestSummary"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ManifestSummary:
      type: object
      properties:
        schemaVersion:
          type: string
        adminUrl:
          type: string
        exportedAt:
          type: string
...
    MigrationPlan:
      type: object
      properties:
        id:
          type: string
        gatewayStrategy:
          type: string
        sourceProducts:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedResource"
        aiAnalysis:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        catalogInfoYaml:
          type: string
        status:
          type: string
        targetClusterId:
          type: string
        targetClusterLabel:
          type: string
        consolidationWarnings:
          type: array
          items:
            type: string
        prerequisites:
          type: array
          items:
            $ref: "#/components/schemas/MigrationPrerequisite"
    MigrationPrerequisite:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        title:
          type: string
        description:
          type: string
        requiredByPlan:
          type: boolean
        optionalTier:
          type: boolean
        docUrl:
          type: string
        status:
          type: string
        triggeredByCount:
          type: integer
          format: int32
...
    ProductSummary:
      type: object
      properties:
        name:
          type: string
        systemName:
          type: string
        serviceId:
          type: integer
          format: int64
...
    ResourceResult:
      type: object
      properties:
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        success:
          type: boolean
        message:
          type: string
...
    TestCommand:
      description: Suggested curl command to validate a migrated route
      type: object
      properties:
        label:
          type: string
        command:
          type: string
        type:
          type: string
          description: "no-auth, bearer, api-key, or path-test"
...
    ThreeScaleSource:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        accessToken:
          type: string
        enabled:
          type: boolean
...
    ThreeScaleSourceView:
      description: 3scale Admin API source (public view — credentials never exposed)
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        adminUrl:
          type: string
        enabled:
          type: boolean
        credentialConfigured:
          type: boolean
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /api/migration/analyze:
    post:
      operationId: analyzeMigration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationPlan"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Analyze
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - operator
        - admin
  /api/migration/import-export:
    post:
      operationId: importExport
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportExportResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Import Export
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - operator
        - admin
...
  /api/migration/plans/{id}:
    get:
      operationId: getPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MigrationPlan"
      summary: Get Plan
      tags:
      - Migration Resource
  /api/migration/plans/{id}/apply:
    post:
      operationId: applyPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyResult"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Apply Plan
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - admin
  /api/migration/plans/{id}/catalog-info/{productName}:
    get:
      operationId: getCatalogInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: productName
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            text/yaml:
              schema:
                type: string
      summary: Get Catalog Info
      tags:
      - Migration Resource
  /api/migration/plans/{id}/confirm-registration:
    post:
      operationId: confirmRegistration
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRegistrationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Confirm Registration
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - admin
  /api/migration/plans/{id}/drift:
    get:
      operationId: checkDrift
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriftEntry"
      summary: Check Drift
      tags:
      - Migration Resource
  /api/migration/plans/{id}/revert:
    post:
      operationId: revertPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyResult"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Revert Plan
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - admin
  /api/migration/plans/{id}/test-commands:
    get:
      operationId: getTestCommands
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCommand"
      summary: Get Test Commands
      tags:
      - Migration Resource
...
  /api/migration/revert-bulk:
    post:
      operationId: revertBulk
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkRevertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkRevertResult"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Revert Bulk
      tags:
      - Migration Resource
      security:
      - SecurityScheme:
        - admin
...
  /api/threescale/sources:
    get:
      operationId: listSources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreeScaleSourceView"
      summary: List Sources
      tags:
      - Three Scale Resource
    post:
      operationId: addSource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThreeScaleSource"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeScaleSourceView"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Add Source
      tags:
      - Three Scale Resource
      security:
      - SecurityScheme:
        - admin
  /api/threescale/sources/{id}:
    delete:
      operationId: removeSource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Remove Source
      tags:
      - Three Scale Resource
      security:
      - SecurityScheme:
        - admin
...
info:
  title: ApiShift API
  version: 0.3.0
  description: 3scale to Connectivity Link migration console REST API
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value